+++ examples/full_example/lib/customs/custom_addon.dart
@@ -3,6 +3,9 @@ import 'package:widgetbook/widgetbook.dart';
 
 class AlignMode extends Mode<Alignment> {
   AlignMode(Alignment value) : super(value, AlignmentAddon(value));
+
+  @override
+  String get formattedValue => AlignmentAddon.alignments[value]!;
 }
 
 /// An [Addon] for wrapping use-cases with [Align] widget.