+++ packages/widgetbook/lib/src/core/addons/alignment_addon.dart
@@ -5,6 +5,9 @@ import '../framework/framework.dart';
 
 class AlignmentMode extends Mode<Alignment> {
   AlignmentMode(Alignment value) : super(value, AlignmentAddon(value));
+
+  @override
+  String get formattedValue => AlignmentAddon.alignments[value]!;
 }
 
 /// An [Addon] for wrapping use-cases with [Align] widget.