+++ packages/widgetbook/lib/src/core/addons/semantics_addon/semantics_addon.dart
@@ -6,6 +6,9 @@ import 'minimal_semantics_debugger.dart';
 
 class SemanticsMode extends Mode<bool> {
   SemanticsMode(bool value) : super(value, SemanticsAddon());
+
+  @override
+  String get formattedValue => value ? 'Enabled' : 'Disabled';
 }
 
 /// An [Addon] for semantics.