+++ packages/widgetbook/lib/src/core/addons/text_scale_addon.dart
@@ -7,6 +7,9 @@ import '../framework/framework.dart';
 /// via [MediaQuery].
 class TextScaleMode extends Mode<double> {
   TextScaleMode(double value) : super(value, TextScaleAddon());
+
+  @override
+  String get formattedValue => 'x$value';
 }
 
 class TextScaleAddon extends Addon<double> with SingleFieldOnly {