+++ packages/widgetbook/lib/src/core/framework/mode.dart
@@ -12,6 +12,11 @@ abstract class Mode<T> {
 
   String get groupName => addon.groupName;
 
+  String get name => addon.name;
+
+  /// The name of this [value] to be used for displaying purposes.
+  String get formattedValue;
+
   /// The associated addon for this mode. Used to build the story, and
   /// to convert to/from query parameters.
   final Addon<T> addon;