+++ packages/widgetbook/lib/src/core/layout/mobile_layout.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
 import '../settings/settings.dart';
+import '../state/widgetbook_state.dart';
 import 'base_layout.dart';
 
 /// The [MobileLayout] is a layout for mobile devices that allows
@@ -13,16 +14,20 @@ class MobileLayout extends StatelessWidget implements BaseLayout {
     required this.navigationBuilder,
     required this.addonsBuilder,
     required this.argsBuilder,
+    required this.scenarioInfoBuilder,
     required this.workbench,
   });
 
   final Widget Function(BuildContext context) navigationBuilder;
   final List<Widget> Function(BuildContext context) addonsBuilder;
   final List<Widget> Function(BuildContext context) argsBuilder;
+  final Widget Function(BuildContext context) scenarioInfoBuilder;
   final Widget workbench;
 
   @override
   Widget build(BuildContext context) {
+    final isScenario = WidgetbookState.of(context).scenario != null;
+
     return Scaffold(
       body: SafeArea(
         child: workbench,
@@ -33,25 +38,33 @@ class MobileLayout extends StatelessWidget implements BaseLayout {
             label: 'Navigation',
             icon: Icon(Icons.list_outlined),
           ),
-          const BottomNavigationBarItem(
-            label: 'Addons',
-            icon: Icon(Icons.dashboard_customize_outlined),
-          ),
-          const BottomNavigationBarItem(
-            label: 'Args',
-            icon: Icon(Icons.tune_outlined),
-          ),
+          if (isScenario) ...{
+            const BottomNavigationBarItem(
+              label: 'Scenario Info',
+              icon: Icon(Icons.info_outline),
+            ),
+          } else ...{
+            const BottomNavigationBarItem(
+              label: 'Addons',
+              icon: Icon(Icons.dashboard_customize_outlined),
+            ),
+            const BottomNavigationBarItem(
+              label: 'Args',
+              icon: Icon(Icons.tune_outlined),
+            ),
+          },
         ],
         onTap: (index) {
           showModalBottomSheet<void>(
             context: context,
             builder: (context) => switch (index) {
               0 => navigationBuilder(context),
-              1 => SettingsList(
+              1 when isScenario => scenarioInfoBuilder(context),
+              1 when !isScenario => SettingsList(
                 name: 'Addons',
                 builder: addonsBuilder,
               ),
-              2 => SettingsList(
+              2 when !isScenario => SettingsList(
                 name: 'Args',
                 builder: argsBuilder,
               ),